fix string

+++ src/error.rs
@@ -1,8 +1,9 @@
-pub fn error(line: i32, message: String){
+
+pub fn error(line: usize, message: String){
     report(line, "".to_string(),message, &mut false);
 }
 
-pub fn report(line: i32, error_where:String, message:String,hasError: &mut bool){
+pub fn report(line: usize, error_where:String, message:String,hasError: &mut bool){
 
     println!("[line {line}] Error {error_where} => {message} ");
     *hasError = true;
